requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
cacheout==0.14.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
fastapi==0.104.0
fastapi-versioning==0.10.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
trio==0.22.2
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
